refactor AdventureBoss SeasonInfo state and handler
